% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhecodeX_definitions.R
\docType{data}
\name{PhecodeX_definitions}
\alias{PhecodeX_definitions}
\title{PhecodeX definitions}
\format{
A data frame with 3612 obs. and 7 variables:
\describe{
  \item{phecode}{Phecode code}
  \item{phecode_string}{Description of phecode}
  \item{category_num}{Number index of phecode category}
  \item{category}{Category of phecode based on ICD body systems}
  \item{sex}{gender constraint for each phecode (Male, Female, or Both)}
  \item{icd10_only}{Does the phecode only consist of ICD10 codes 1=Y, 0=N}
  \item{phecode_num}{Phecode code numeric version}
}
}
\source{
\url{https://phewascatalog.org/files/phecodeX/phecodeX_info.csv.zip}
}
\usage{
PhecodeX_definitions
}
\description{
A dataset containing PhecodeX diagnosis categories and their descriptions
}
\seealso{
\code{\link{PhecodeX_map}} for ICD code mapping for each Phecode.
}
\keyword{datasets}
